+++ contracts/contract/src/lib.rs
@@ -34,6 +34,7 @@ mod actions_pyth;
 mod protocol_debts;
 mod storage_keys;
 mod client_echo;
+mod booster_tokens;
 
 pub use crate::account::*;
 pub use crate::account_asset::*;
@@ -67,6 +68,7 @@ pub use crate::pyth::*;
 pub use crate::protocol_debts::*;
 pub use crate::storage_keys::*;
 pub use crate::client_echo::*;
+pub use crate::booster_tokens::*;
 
 use common::*;
 
@@ -80,7 +82,7 @@ use near_sdk::{
     Duration, Gas, PanicOnDefault, Promise, Timestamp, require, promise_result_as_success
 };
 use once_cell::sync::Lazy;
-use std::collections::HashMap;
+use std::collections::{HashMap, HashSet};
 use std::sync::Mutex;
 
 #[derive(BorshSerialize, BorshStorageKey)]
@@ -307,6 +309,11 @@ mod unit_env {
     }
 
     impl UnitEnv {
+        pub fn init_booster_tokens(&mut self){
+            testing_env!(self.context.predecessor_account_id(owner_id()).attached_deposit(1).build());
+            self.contract.add_booster_token_info(booster_token_id(), 18, 2678400, 31536000, 40000, U128(1));
+        }
+
         pub fn init_users(&mut self){
             testing_env!(self.context.predecessor_account_id(owner_id()).attached_deposit(d(1, 23)).build());
             self.contract.storage_deposit(Some(alice()), None);
@@ -562,7 +569,7 @@ mod unit_env {
             reward_amount: Balance
         ){
             testing_env!(self.context.predecessor_account_id(owner_id()).attached_deposit(1).build());
-            self.contract.add_asset_farm_reward(farm_id, reward_token_id, new_reward_per_day.into(), new_booster_log_base.into(), reward_amount.into());
+            self.contract.add_asset_farm_reward(farm_id, reward_token_id, new_reward_per_day.into(), HashMap::from([(booster_token_id(), U128(new_booster_log_base))]), reward_amount.into());
         }
 
         pub fn account_farm_claim_all(&mut self, account_id: AccountId){
@@ -692,6 +699,7 @@ mod unit_env {
         value * 10u128.pow(decimals as _)
     }
 
+    #[allow(deprecated)]
     pub fn init_unit_env() -> UnitEnv {
         let mut context = VMContextBuilder::new();
         testing_env!(context.predecessor_account_id(owner_id()).build());
@@ -722,6 +730,7 @@ mod unit_env {
         };
         test_env.init_assets();
         test_env.init_users();
+        test_env.init_booster_tokens();
         test_env
     }
 
@@ -925,19 +934,19 @@ mod booster {
         let account = test_env.contract.get_account(alice()).unwrap();
         assert_eq!(account.supplied[0].balance, amount);
         assert_eq!(account.supplied[0].token_id, booster_token_id());
-        assert!(account.booster_staking.is_none());
+        assert!(account.booster_stakings.is_empty());
 
         let duration_sec: DurationSec = MAX_DURATION_SEC;
 
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_stake_booster(Some(amount.into()), duration_sec);
+        test_env.contract.account_stake_booster(booster_token_id(), Some(amount.into()), duration_sec);
         
         let asset = test_env.get_asset(booster_token_id());
         assert_eq!(asset.supplied.balance, 0);
 
         let account = test_env.contract.get_account(alice()).unwrap();
         assert!(account.supplied.is_empty());
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.staked_booster_amount, amount);
         assert_eq!(booster_staking.x_booster_amount, amount * 4);
         assert_eq!(
@@ -949,18 +958,18 @@ mod booster {
 
         let account = test_env.contract.get_account(alice()).unwrap();
         assert!(account.supplied.is_empty());
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.x_booster_amount, amount * 4);
 
         test_env.skip_time_to_by_sec(duration_sec);
 
         let account = test_env.contract.get_account(alice()).unwrap();
         assert!(account.supplied.is_empty());
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.x_booster_amount, amount * 4);
 
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_unstake_booster();
+        test_env.contract.account_unstake_booster(Some(booster_token_id()), None);
 
         let asset = test_env.get_asset(booster_token_id());
         assert_eq!(asset.supplied.balance, amount);
@@ -968,7 +977,7 @@ mod booster {
         let account = test_env.contract.get_account(alice()).unwrap();
         assert_eq!(account.supplied[0].balance, amount);
         assert_eq!(account.supplied[0].token_id, booster_token_id());
-        assert!(account.booster_staking.is_none());
+        assert!(account.booster_stakings.is_empty());
     }
 
 
@@ -983,7 +992,7 @@ mod booster {
         let duration_sec: DurationSec = MAX_DURATION_SEC;
 
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_stake_booster(Some((amount / 2).into()), duration_sec);
+        test_env.contract.account_stake_booster(booster_token_id(), Some((amount / 2).into()), duration_sec);
 
         let asset = test_env.get_asset(booster_token_id());
         assert_eq!(asset.supplied.balance, amount / 2);
@@ -991,7 +1000,7 @@ mod booster {
         let account = test_env.contract.get_account(alice()).unwrap();
         assert_eq!(account.supplied[0].balance, amount / 2);
         assert_eq!(account.supplied[0].token_id, booster_token_id());
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.staked_booster_amount, amount / 2);
         assert_eq!(booster_staking.x_booster_amount, amount / 2 * 4);
         assert_eq!(
@@ -1002,14 +1011,14 @@ mod booster {
         test_env.skip_time_to_by_sec(duration_sec / 2 );
 
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_stake_booster(Some((amount / 2).into()), duration_sec / 2);
+        test_env.contract.account_stake_booster(booster_token_id(), Some((amount / 2).into()), duration_sec / 2);
 
         let asset = test_env.get_asset(booster_token_id());
         assert_eq!(asset.supplied.balance, 0);
 
         let account = test_env.contract.get_account(alice()).unwrap();
         assert!(account.supplied.is_empty());
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.staked_booster_amount, amount);
         assert_eq!(
             booster_staking.x_booster_amount,
@@ -1028,7 +1037,7 @@ mod booster {
         test_env.skip_time_to_by_sec(duration_sec);
 
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_unstake_booster();
+        test_env.contract.account_unstake_booster(Some(booster_token_id()), None);
 
         let asset = test_env.get_asset(booster_token_id());
         assert_eq!(asset.supplied.balance, amount);
@@ -1036,7 +1045,7 @@ mod booster {
         let account = test_env.contract.get_account(alice()).unwrap();
         assert_eq!(account.supplied[0].balance, amount);
         assert_eq!(account.supplied[0].token_id, booster_token_id());
-        assert!(account.booster_staking.is_none());
+        assert!(account.booster_stakings.is_empty());
     }
 
     #[test]
@@ -1050,19 +1059,19 @@ mod booster {
         let duration_sec: DurationSec = MAX_DURATION_SEC;
 
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_stake_booster(Some((amount / 2).into()), duration_sec);
+        test_env.contract.account_stake_booster(booster_token_id(), Some((amount / 2).into()), duration_sec);
         
         test_env.skip_time_to_by_sec(duration_sec / 2 );
 
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_stake_booster(Some((amount / 2).into()), duration_sec);
+        test_env.contract.account_stake_booster(booster_token_id(), Some((amount / 2).into()), duration_sec);
 
         let asset = test_env.get_asset(booster_token_id());
         assert_eq!(asset.supplied.balance, 0);
 
         let account = test_env.contract.get_account(alice()).unwrap();
         assert!(account.supplied.is_empty());
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.staked_booster_amount, amount);
         assert_eq!(booster_staking.x_booster_amount, amount * 4);
         assert_eq!(
@@ -1071,7 +1080,7 @@ mod booster {
         );
 
         test_env.skip_time_to_by_sec(duration_sec + duration_sec / 2);
-        test_env.contract.account_unstake_booster();
+        test_env.contract.account_unstake_booster(Some(booster_token_id()), None);
 
         let asset = test_env.get_asset(booster_token_id());
         assert_eq!(asset.supplied.balance, amount);
@@ -1079,7 +1088,7 @@ mod booster {
         let account = test_env.contract.get_account(alice()).unwrap();
         assert_eq!(account.supplied[0].balance, amount);
         assert_eq!(account.supplied[0].token_id, booster_token_id());
-        assert!(account.booster_staking.is_none());
+        assert!(account.booster_stakings.is_empty());
     }
 }
 
@@ -1364,7 +1373,7 @@ mod farms {
         test_env.deposit(booster_token_id(), alice(), booster_amount);
         
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_stake_booster(Some(booster_amount.into()), MAX_DURATION_SEC);
+        test_env.contract.account_stake_booster(booster_token_id(), Some(booster_amount.into()), MAX_DURATION_SEC);
 
         let reward_per_day = d(100, 18);
         let total_reward = d(3000, 18);
@@ -1378,29 +1387,29 @@ mod farms {
         test_env.account_farm_claim_all(alice());
 
         let account = test_env.contract.get_account(alice()).unwrap();
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.staked_booster_amount, d(5, 18));
         assert_eq!(booster_staking.x_booster_amount, d(20, 18));
         assert_eq!(booster_staking.unlock_timestamp, to_nano(10 + MAX_DURATION_SEC));
         assert_eq!(account.farms[0].rewards[0].boosted_shares, d(200, 18));
         
         testing_env!(test_env.context.predecessor_account_id(owner_id()).build());
-        test_env.contract.adjust_boost_staking_policy(2678400, MAX_DURATION_SEC / 2, 40000);
+        test_env.contract.update_booster_token_info(booster_token_id(), Some((2678400, MAX_DURATION_SEC / 2)), Some(40000), None, None);
         test_env.account_farm_claim_all(alice());
         
         let account = test_env.contract.get_account(alice()).unwrap();
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.staked_booster_amount, d(5, 18));
         assert_eq!(booster_staking.x_booster_amount, d(20, 18));
         assert_eq!(booster_staking.unlock_timestamp, to_nano(10 + MAX_DURATION_SEC / 2));
         assert_eq!(account.farms[0].rewards[0].boosted_shares, d(200, 18));
 
         testing_env!(test_env.context.predecessor_account_id(owner_id()).build());
-        test_env.contract.adjust_boost_staking_policy(2678400, MAX_DURATION_SEC / 2, 20000);
+        test_env.contract.update_booster_token_info(booster_token_id(), Some((2678400, MAX_DURATION_SEC / 2)), Some(20000), None, None);
         test_env.account_farm_claim_all(alice());
 
         let account = test_env.contract.get_account(alice()).unwrap();
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.staked_booster_amount, d(5, 18));
         assert_eq!(booster_staking.x_booster_amount, d(10, 18));
         assert_eq!(booster_staking.unlock_timestamp, to_nano(10 + MAX_DURATION_SEC / 2));
@@ -1411,7 +1420,7 @@ mod farms {
 
         test_env.account_farm_claim_all(alice());
         let account = test_env.contract.get_account(alice()).unwrap();
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.staked_booster_amount, d(5, 18));
         assert_eq!(booster_staking.x_booster_amount, 0);
         assert_eq!(booster_staking.unlock_timestamp, to_nano(10 + MAX_DURATION_SEC / 2));
@@ -1429,7 +1438,7 @@ mod farms {
 
         test_env.account_farm_claim_all(alice());
         let account = test_env.contract.get_account(alice()).unwrap();
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.staked_booster_amount, d(5, 18));
         assert_eq!(booster_staking.x_booster_amount, 0);
         assert_eq!(booster_staking.unlock_timestamp, to_nano(10 + MAX_DURATION_SEC / 2));
@@ -1450,7 +1459,7 @@ mod farms {
         test_env.deposit(booster_token_id(), alice(), booster_amount);
         
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_stake_booster(Some(booster_amount.into()), MAX_DURATION_SEC);
+        test_env.contract.account_stake_booster(booster_token_id(), Some(booster_amount.into()), MAX_DURATION_SEC);
 
         let reward_per_day = d(100, 18);
         let total_reward = d(3000, 18);
@@ -1464,7 +1473,7 @@ mod farms {
         test_env.account_farm_claim_all(alice());
 
         let account = test_env.contract.get_account(alice()).unwrap();
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.staked_booster_amount, d(5000, 18));
         assert_eq!(booster_staking.x_booster_amount, d(20000, 18));
         assert_eq!(booster_staking.unlock_timestamp, to_nano(10 + MAX_DURATION_SEC));
@@ -1472,10 +1481,10 @@ mod farms {
             * ((d(20000, 18) as f64) / (10f64.powi(18))).log(20f64)) as u128);
 
         testing_env!(test_env.context.predecessor_account_id(owner_id()).build());
-        test_env.contract.adjust_boost_suppress_factor(1000);
+        test_env.contract.update_booster_token_info(booster_token_id(), None, None, Some(U128(1000)), None);
         test_env.account_farm_claim_all(alice());
         let account = test_env.contract.get_account(alice()).unwrap();
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.staked_booster_amount, d(5000, 18));
         assert_eq!(booster_staking.x_booster_amount, d(20000, 18));
         assert_eq!(booster_staking.unlock_timestamp, to_nano(10 + MAX_DURATION_SEC));
@@ -1492,7 +1501,8 @@ mod farms {
         let mut test_env = init_unit_env();
 
         testing_env!(test_env.context.predecessor_account_id(owner_id()).attached_deposit(1).build());
-        test_env.contract.adjust_boost_staking_policy(MIN_DURATION_SEC, MAX_DURATION_SEC, 120000);
+        test_env.contract.update_booster_token_info(booster_token_id(), Some((MIN_DURATION_SEC, MAX_DURATION_SEC)), Some(120000), None, None);
+
 
         test_env.skip_time_to_by_sec(10);
 
@@ -1500,10 +1510,10 @@ mod farms {
         test_env.deposit(booster_token_id(), alice(), booster_amount);
         
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_stake_booster(Some(d(5, 18).into()), MAX_DURATION_SEC);
+        test_env.contract.account_stake_booster(booster_token_id(), Some(d(5, 18).into()), MAX_DURATION_SEC);
 
         let account = test_env.contract.get_account(alice()).unwrap();
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.staked_booster_amount, d(5, 18));
         assert_eq!(booster_staking.x_booster_amount, d(60, 18));
         assert_eq!(booster_staking.unlock_timestamp, to_nano(10 + MAX_DURATION_SEC));
@@ -1511,20 +1521,21 @@ mod farms {
         test_env.skip_time_to_by_sec(10 + MIN_DURATION_SEC * 2);
 
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_stake_booster(Some(d(5, 18).into()), MAX_DURATION_SEC);
+        test_env.contract.account_stake_booster(booster_token_id(), Some(d(5, 18).into()), MAX_DURATION_SEC);
         let account = test_env.contract.get_account(alice()).unwrap();
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.staked_booster_amount, d(10, 18));
         assert_eq!(booster_staking.x_booster_amount, d(120, 18));
         assert_eq!(booster_staking.unlock_timestamp, to_nano(10 + MIN_DURATION_SEC * 2 + MAX_DURATION_SEC));
 
         testing_env!(test_env.context.predecessor_account_id(owner_id()).build());
-        test_env.contract.adjust_boost_staking_policy(MIN_DURATION_SEC, MIN_DURATION_SEC * 3, 30000);
+        test_env.contract.update_booster_token_info(booster_token_id(), Some((MIN_DURATION_SEC, MIN_DURATION_SEC * 3)), Some(30000), None, None);
+
 
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_stake_booster(Some(d(5, 18).into()), MIN_DURATION_SEC * 3);
+        test_env.contract.account_stake_booster(booster_token_id(), Some(d(5, 18).into()), MIN_DURATION_SEC * 3);
         let account = test_env.contract.get_account(alice()).unwrap();
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.staked_booster_amount, d(15, 18));
         assert_eq!(booster_staking.x_booster_amount, d(45, 18));
         assert_eq!(booster_staking.unlock_timestamp, to_nano(10 + MIN_DURATION_SEC * 5));
@@ -1532,12 +1543,12 @@ mod farms {
         test_env.skip_time_to_by_sec(10 + MIN_DURATION_SEC * 4);
 
         testing_env!(test_env.context.predecessor_account_id(owner_id()).build());
-        test_env.contract.adjust_boost_staking_policy(MIN_DURATION_SEC, MAX_DURATION_SEC, 120000);
+        test_env.contract.update_booster_token_info(booster_token_id(), Some((MIN_DURATION_SEC, MAX_DURATION_SEC)), Some(120000), None, None);
 
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_stake_booster(Some(d(5, 18).into()), MAX_DURATION_SEC);
+        test_env.contract.account_stake_booster(booster_token_id(), Some(d(5, 18).into()), MAX_DURATION_SEC);
         let account = test_env.contract.get_account(alice()).unwrap();
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.staked_booster_amount, d(20, 18));
         assert_eq!(booster_staking.x_booster_amount, d(240, 18));
         assert_eq!(booster_staking.unlock_timestamp, to_nano(10 + MIN_DURATION_SEC * 4 + MAX_DURATION_SEC));
@@ -1564,7 +1575,7 @@ mod farms {
         test_env.deposit(booster_token_id(), alice(), booster_amount);
 
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_stake_booster(Some(booster_amount.into()), MAX_DURATION_SEC);
+        test_env.contract.account_stake_booster(booster_token_id(), Some(booster_amount.into()), MAX_DURATION_SEC);
     
         let amount = d(100, 18);
         test_env.deposit(ndai_token_id(), alice(), amount);
@@ -1585,7 +1596,7 @@ mod farms {
         let account = test_env.contract.get_account(alice()).unwrap();
         assert_balances(&account.supplied, &[av(ndai_token_id(), amount)]);
 
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.staked_booster_amount, booster_amount);
         assert_eq!(booster_staking.x_booster_amount, booster_amount * 4);
 
@@ -1621,7 +1632,7 @@ mod farms {
 
         // Increasing booster stake updates all farms.
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_stake_booster(Some(booster_amount.into()), MAX_DURATION_SEC);
+        test_env.contract.account_stake_booster(booster_token_id(), Some(booster_amount.into()), MAX_DURATION_SEC);
 
         let asset = test_env.get_asset(nusdc_token_id());
         assert_eq!(asset.supplied.balance, farmed_amount);
@@ -1656,7 +1667,7 @@ mod farms {
                 * 3,
         );
         assert_eq!(account.farms[0].rewards[0].unclaimed_amount, 0);
-        let booster_staking = account.booster_staking.unwrap();
+        let booster_staking = account.booster_stakings.get(&booster_token_id()).unwrap();
         assert_eq!(booster_staking.staked_booster_amount, d(100, 18));
         assert_eq!(booster_staking.x_booster_amount, d(400, 18));
         clean_assets_cache();
@@ -1674,7 +1685,7 @@ mod farms {
         test_env.deposit(booster_token_id(), alice(), booster_amount);
 
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_stake_booster(Some(booster_amount.into()), MAX_DURATION_SEC);
+        test_env.contract.account_stake_booster(booster_token_id(), Some(booster_amount.into()), MAX_DURATION_SEC);
 
         test_env.skip_time_to_by_sec(10 + MAX_DURATION_SEC - ONE_DAY_SEC * 3);
 
@@ -1732,7 +1743,7 @@ mod farms {
 
         // Unstaking booster updates all farms.
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_unstake_booster();
+        test_env.contract.account_unstake_booster(Some(booster_token_id()), None);
 
         let asset = test_env.get_asset(nusdc_token_id());
         assert_eq!(asset.supplied.balance, farmed_amount);
@@ -1767,7 +1778,7 @@ mod farms {
                 .0,
         );
         assert_eq!(account.farms[0].rewards[0].unclaimed_amount, 0);
-        assert!(account.booster_staking.is_none());
+        assert!(account.booster_stakings.get(&booster_token_id()).is_none());
         clean_assets_cache();
         clean_assets_farm_cache();
     }
@@ -1783,13 +1794,13 @@ mod farms {
         test_env.deposit(booster_token_id(), alice(), booster_amount_alice);
 
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_stake_booster(Some(booster_amount_alice.into()), MAX_DURATION_SEC);
+        test_env.contract.account_stake_booster(booster_token_id(), Some(booster_amount_alice.into()), MAX_DURATION_SEC);
 
         let booster_amount_bob = d(100, 18);
         test_env.deposit(booster_token_id(), bob(), booster_amount_bob);
 
         testing_env!(test_env.context.predecessor_account_id(bob()).attached_deposit(1).build());
-        test_env.contract.account_stake_booster(Some(booster_amount_bob.into()), MAX_DURATION_SEC);
+        test_env.contract.account_stake_booster(booster_token_id(), Some(booster_amount_bob.into()), MAX_DURATION_SEC);
 
         let reward_per_day = d(100, 18);
         let total_reward = d(3000, 18);
@@ -1872,7 +1883,7 @@ mod farms {
 
         // Increasing booster stake updates all farms.
         testing_env!(test_env.context.predecessor_account_id(alice()).attached_deposit(1).build());
-        test_env.contract.account_stake_booster(Some(extra_booster_amount.into()), MAX_DURATION_SEC);
+        test_env.contract.account_stake_booster(booster_token_id(), Some(extra_booster_amount.into()), MAX_DURATION_SEC);
 
         let asset = test_env.get_asset(nusdc_token_id());
         // The amount of only for Alice, but Bob still unclaimed
@@ -2786,5 +2797,42 @@ mod liquidation {
         test_env.contract.storage.insert(&AccountId::new_unchecked("storage".to_string()), &VStorage::Current(Storage { storage_balance: 10u128.pow(25), used_bytes: 1000, storage_tracker: Default::default() }));
         let _tmp_account = account.clone();
         test_env.contract.internal_set_account(&AccountId::new_unchecked("storage".to_string()), account);
+    }
+
+    #[test]
+    fn test_get_booster_extra_shares() {
+        let user_account_id = AccountId::new_unchecked("user".to_string());
+        let booster_token_id = AccountId::new_unchecked("booster".to_string());
+        let booster_token_id2 = AccountId::new_unchecked("booster2".to_string());
+        let mut account = Account::new(&user_account_id);
+        let mut asset_farm_reward: AssetFarmReward = Default::default();
+        let mut booster_tokens: HashMap<TokenId, BoosterTokenInfo> = Default::default();
+        assert_eq!(get_booster_extra_shares(&account, 10000, &asset_farm_reward, &booster_tokens), 0);
+        booster_tokens.insert(booster_token_id.clone(), BoosterTokenInfo::new(
+            booster_token_id.clone(),
+            18,
+            2678400,
+            31536000,
+            40000,
+            1
+        ));
+        asset_farm_reward.booster_log_bases.insert(booster_token_id.clone(), 10u128.pow(19).into());
+        assert_eq!(get_booster_extra_shares(&account, 10000, &asset_farm_reward, &booster_tokens), 0);
+        account.booster_stakings.insert(booster_token_id.clone(), BoosterStaking { staked_booster_amount: 50, x_booster_amount: 100, unlock_timestamp: u64::MAX });
+        assert_eq!(get_booster_extra_shares(&account, 10000, &asset_farm_reward, &booster_tokens), 0);
+        account.booster_stakings.insert(booster_token_id.clone(), BoosterStaking { staked_booster_amount: 50 * 10u128.pow(18), x_booster_amount: 100 * 10u128.pow(18), unlock_timestamp: u64::MAX });
+        assert_eq!(get_booster_extra_shares(&account, 10000, &asset_farm_reward, &booster_tokens), 20000);
+        booster_tokens.insert(booster_token_id2.clone(), BoosterTokenInfo::new(
+            booster_token_id2.clone(),
+            18,
+            2678400,
+            31536000,
+            40000,
+            1
+        ));
+        account.booster_stakings.insert(booster_token_id2.clone(), BoosterStaking { staked_booster_amount: 50 * 10u128.pow(18), x_booster_amount: 100 * 10u128.pow(18), unlock_timestamp: u64::MAX });
+        assert_eq!(get_booster_extra_shares(&account, 10000, &asset_farm_reward, &booster_tokens), 20000);
+        asset_farm_reward.booster_log_bases.insert(booster_token_id2.clone(), 10u128.pow(19).into());
+        assert_eq!(get_booster_extra_shares(&account, 10000, &asset_farm_reward, &booster_tokens), 40000);
     }
 }