+++ contracts/contract/src/storage_keys.rs
@@ -2,3 +2,4 @@ pub const PROTOCOL_DEBTS_KEY: &str = "pd";
 pub const MARGIN_BASE_TOKEN_LIMIT: &str = "bl";
 pub const MARGIN_BASE_TOKEN_LIMIT_GUR: &str = "bg";
 pub const CLIENT_ECHO_SENDER_WHITELIST: &str = "csw";
+pub const BOOSTER_TOKENS_KEY: &str = "bt";